Add bird animation

+++ src/Clock.java
@@ -2,13 +2,13 @@ import java.util.*;
 import java.util.function.*;
 
 public class Clock {
-    private List<Function<Double, Void>> functions = new ArrayList<Function<Double, Void>>();
+    private List<Function<Float, Void>> functions = new ArrayList<Function<Float, Void>>();
 
-    public void register(Function<Double, Void> function) {
+    public void register(Function<Float, Void> function) {
         functions.add(function);
     }
 
-    public void elapse(double time) {
+    public void elapse(float time) {
         functions.forEach(function -> {
             function.apply(time);
         });
